experiments/pyproject.toml
[tool.poetry]
name = "pmlib"
version = "1.2.0"
description = "Description of pmlib"
authors = ["Jan Lukany <jan.lukany@datamole.cz>"]
exclude = ["tests"]

[[tool.poetry.source]]
name = "dtml"
url = "https://pypi.datamole.cz/packages/release/+simple"

[tool.poetry.dependencies]
python = "^3.7"
matplotlib = ">=3.1"
numpy = ">=1.18"
pandas = ">=1.0.1"
scikit-learn = ">=0.22.1"

[tool.poetry.dev-dependencies]
seaborn = "^0.10.0"
pyarrow = "^0.16.0"
pyprg = "^0.1.1-beta.7"
autopep8 = ">=1.4"
ipywidgets = ">=7.5.1"
jupyterlab = ">=1.2.6"
jupytext = ">=1.3.3"
flake8 = ">=3.7.9"
pep8 = ">=1.7"
tabulate = ">=0.8.6"
pytest = ">=5.3"
pytest-cov = ">=2.8"
rope = ">=0.16.0"
sphinx = ">=2.3"
sphinx-rtd-theme = ">=0.4.3"
sphinxcontrib-bibtex = ">=1.0.0"
tqdm = "*"
xgboost = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
